move cartopy helper functions to standalone file
